fix: format model prices with registered units
